Refactor: Break down ReceiptPageMapper into smaller services
